services/agents/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "curation-persona-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "fastapi",
    "uvicorn[standard]",
    "google-cloud-firestore",
    "google-genai",
    "feedparser",
    "python-dateutil",
    "httpx",
    "pydantic>=2.0",
    "pydantic-settings",
    "beautifulsoup4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
]

[tool.setuptools.packages.find]
include = ["shared*", "collector*", "librarian*", "researcher*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
